+++ FindPOLine/utils.js
@@ -128,3 +128,57 @@ export function getLinesQuery(queryParams, ky, localeDateFormat, customFields) {
     return buildOrderLinesQuery(queryParams, isbnData?.isbn, isbnData?.isbnType, localeDateFormat, customFields);
   };
 }
+
+/**
+ * Generate options with deprecated labels for a dropdown.
+ * @param {object[]} records - array of json objects with { name, deprecated }
+ * @param {class} intl - class for internationalization
+ * @param {string} deprecatedMessageId - intl message ID for labels
+ * @returns {object[]} array of {label, value} pairs.
+ */
+const generateOptionsWithDeprecationLabels = (
+  records,
+  intl,
+  deprecatedMessageId,
+) => records
+  .map(({ name, deprecated }) => {
+    return {
+      label: deprecated
+        ? intl.formatMessage({ id: deprecatedMessageId }, { name })
+        : name,
+      value: name,
+    };
+  });
+
+/**
+ * Calculate the options of a prefix to use in a dropdown.
+ * https://github.com/folio-org/acq-models/blob/master/mod-orders-storage/schemas/prefix.json
+ * @param {object[]} records - array of json objects (prefix) with { name, deprecated }
+ * @param {string} initialSelectedValue - the value of the dropdown on initialization
+ * @param {class} intl - class for internationalization
+ * @returns {object[]} array of {label, value} pairs.
+ */
+export const getPrefixOptions = (
+  records,
+  intl,
+) => generateOptionsWithDeprecationLabels(
+  records,
+  intl,
+  'ui-plugin-find-po-line.filter.prefixFilter.deprecated',
+);
+
+/**
+ * Calculate the options of a suffix to use in a dropdown.
+ * https://github.com/folio-org/acq-models/blob/master/mod-orders-storage/schemas/suffix.json
+ * @param {object[]} records - array of json objects (suffix) with { name, deprecated }
+ * @param {class} intl - class for internationalization
+ * @returns {object[]} array of {label, value} pairs.
+ */
+export const getSuffixOptions = (
+  records,
+  intl,
+) => generateOptionsWithDeprecationLabels(
+  records,
+  intl,
+  'ui-plugin-find-po-line.filter.suffixFilter.deprecated',
+);